remove deprecation warnings from DQMServices/Components
