+++ DQMServices/Components/plugins/DQMScalInfo.cc
@@ -4,18 +4,74 @@
  * Last Update:
  *
  */
+/*
+ * \file DQMDcsInfo.h
+ *
+ * \author A.Meyer - DESY
+ *
+*/
 
-#include "DQMScalInfo.h"
-#include "DataFormats/Scalers/interface/DcsStatus.h"
+#include "DQMServices/Core/interface/DQMEDAnalyzer.h"
+#include "DQMServices/Core/interface/DQMStore.h"
 #include "DataFormats/Common/interface/Handle.h"
-#include "DataFormats/L1GlobalTrigger/interface/L1GtFdlWord.h"
 #include "DataFormats/L1GlobalTrigger/interface/L1GlobalTriggerReadoutRecord.h"
-
-#include "DataFormats/Scalers/interface/Level1TriggerScalers.h"
+#include "DataFormats/L1GlobalTrigger/interface/L1GlobalTriggerReadoutRecord.h"
+#include "DataFormats/L1GlobalTrigger/interface/L1GtFdlWord.h"
+#include "DataFormats/Scalers/interface/DcsStatus.h"
+#include "DataFormats/Scalers/interface/DcsStatus.h"
 #include "DataFormats/Scalers/interface/Level1TriggerRates.h"
+#include "DataFormats/Scalers/interface/Level1TriggerRates.h"
+#include "DataFormats/Scalers/interface/Level1TriggerScalers.h"
+#include "DataFormats/Scalers/interface/Level1TriggerScalers.h"
+#include "DataFormats/Scalers/interface/Level1TriggerScalers.h"
+#include "DataFormats/Scalers/interface/LumiScalers.h"
+#include "DataFormats/Scalers/interface/LumiScalers.h"
+#include "DataFormats/Scalers/interface/ScalersRaw.h"
 #include "DataFormats/Scalers/interface/ScalersRaw.h"
 #include "DataFormats/Scalers/interface/TimeSpec.h"
-#include "DataFormats/Scalers/interface/LumiScalers.h"
+#include "DataFormats/Scalers/interface/TimeSpec.h"
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+#include "FWCore/Framework/interface/Run.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "FWCore/ParameterSet/interface/Registry.h"
+#include "FWCore/ServiceRegistry/interface/Service.h"
+
+class DQMScalInfo : public DQMEDAnalyzer {
+public:
+  /// Constructor
+  DQMScalInfo(const edm::ParameterSet& ps);
+
+  /// Destructor
+  ~DQMScalInfo() override = default;
+
+protected:
+  /// Analyze
+  void analyze(const edm::Event& e, const edm::EventSetup& c) override;
+  void bookHistograms(DQMStore::IBooker&, edm::Run const&, edm::EventSetup const&) override;
+
+private:
+  void makeL1Scalars(const edm::Event& e);
+  void makeLumiScalars(const edm::Event& e);
+
+  edm::ParameterSet parameters_;
+  std::string scalfolder_;
+  edm::EDGetTokenT<L1GlobalTriggerReadoutRecord> gtCollection_;
+  edm::EDGetTokenT<DcsStatusCollection> dcsStatusCollection_;
+  edm::EDGetTokenT<Level1TriggerScalersCollection> l1tscollectionToken_;
+  edm::EDGetTokenT<LumiScalersCollection> lumicollectionToken_;
+
+  // histograms
+  MonitorElement* hlresync_;
+  MonitorElement* hlOC0_;
+  MonitorElement* hlTE_;
+  MonitorElement* hlstart_;
+  MonitorElement* hlEC0_;
+  MonitorElement* hlHR_;
+  MonitorElement* hphysTrig_;
+
+  MonitorElement* hinstLumi_;
+};
 
 using namespace std;
 using namespace edm;
@@ -35,8 +91,6 @@ DQMScalInfo::DQMScalInfo(const edm::ParameterSet& ps) {
   lumicollectionToken_ = consumes<LumiScalersCollection>(
       parameters_.getUntrackedParameter<edm::InputTag>("lumiCollection", edm::InputTag("scalersRawToDigi")));
 }
-
-DQMScalInfo::~DQMScalInfo() = default;
 
 void DQMScalInfo::bookHistograms(DQMStore::IBooker& ibooker,
                                  edm::Run const& /* iRun */,
@@ -109,3 +163,7 @@ void DQMScalInfo::makeLumiScalars(const edm::Event& e) {
 
   return;
 }
+
+#include "FWCore/PluginManager/interface/ModuleDef.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+DEFINE_FWK_MODULE(DQMScalInfo);
